% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_video}
\alias{download_video}
\title{tk_dl_video}
\usage{
download_video(url, path, ua = default_ua)
}
\arguments{
\item{url}{url of tiktok to scrape}

\item{path}{path to download tiktok video to}
}
\description{
Function that enable to download tiktoks
}
