% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_count}
\alias{get_count}
\title{get_count}
\usage{
get_count(
  type,
  cursor = 0,
  ...,
  count = 1,
  save = F,
  path = NULL,
  query = NULL
)
}
\description{
Main function to get data from tiktok
}
