% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(
  url,
  ua = default_ua,
  parse = T,
  port = NULL,
  vpn = F,
  id_cookie = "",
  signed = F
)
}
\arguments{
\item{url}{url to visit and get data from}

\item{parse}{logical. whether to return parsed data or not. Defautls to \code{TRUE}.}
}
\description{
Function for http queries
}
