% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_hashtag}
\alias{get_hashtag}
\title{get_hashtag}
\usage{
get_hashtag(music_id)

get_hashtag(music_id)
}
\description{
Function to get information on a given hashtag

Function to get information on a given piece of music
}
