% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_music_post}
\alias{get_music_post}
\title{get_music_post}
\usage{
<<<<<<< HEAD
<<<<<<< HEAD
get_music_post(music_id, n = 10000, port = NULL, ua = default_ua, ...)
=======
get_music_post(count, music_id, ...)
>>>>>>> 16eb9ca88d97851895eba630e32e62f679ed8cf8
=======
get_music_post(count, music_id, port = NULL, ...)
>>>>>>> 5aea75f85e35f9e94fcecf3b8ebdf3d43888f62d
}
\arguments{
\item{music_id}{Character indicating the \code{music_id} to scrape}

\item{n}{Numeric indicating the number of tiktoks to scrape}
}
\description{
Function to get tiktoks from a given music piece
}
