% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_trending}
\alias{get_trending}
\title{get_trending}
\usage{
get_trending(n = 10000, ua = default_ua)
}
\arguments{
\item{n}{A numeric scalar indicating the number of tiktoks to scrape}
}
\description{
Function to get trending tiktoks
}
