% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json_structure}
\alias{parse_json_structure}
\title{parse_json_structure}
\usage{
parse_json_structure(x)
}
\arguments{
\item{x}{json to be parsed}
}
\description{
parse nested data frame in json responses
}
