% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{tk_auth}
\alias{tk_auth}
\title{Set up a TikTok cookie}
\usage{
tk_auth(cookie, ua, id_cookie)
}
\arguments{
\item{cookie}{The TikTok cookie as string. Find out here how to get a TikTok cookie: \url{https://www.youtube.com/watch?v=kYMV2ugxacs&feature=youtu.be}}
}
\description{
This function will add a TikTok cookie to your
  \code{.Renviron} file so it can be called securely without being stored in
  your code. After you have installed your key, it can be called any time by
  typing \code{Sys.getenv("TIKTOK_COOKIE")}. If you do not have an \code{.Renviron} file, the
  function will create one for you.
}
\examples{

\dontrun{
# Put in your cookie:
tk_auth(cookie = "abcd012345678901234567890123456789")
# Restart R for changes to take effect or load your environment so you can use the cookie without
# restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("TIKTOK_COOKIE")
}
}
