% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tk_discover}
\alias{tk_discover}
\title{tk_discover}
\usage{
tk_discover(scope, save_dir = NULL, port = NULL, ua = default_ua)
}
\arguments{
\item{scope}{Character indicating the endpoint to scrape (must be "hashtag" or "music")}
}
\description{
Function to get 20 suggested hashtags or music_ids
}
