% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tk_dwnl}
\alias{tk_dwnl}
\title{tk_dwnl}
\usage{
tk_dwnl(url, path, time_out = 10)
}
\arguments{
\item{url}{url of tiktok to scrape}

\item{path}{path to download tiktok video to}
}
\description{
Function that enable to download tiktoks
}
