% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tk_dwnl_from_id}
\alias{tk_dwnl_from_id}
\title{tk_dwnl_from_id}
\usage{
tk_dwnl_from_id(post_id = NULL, url = NULL, path, verbose = T, ...)
}
\arguments{
\item{url}{url of tiktok to scrape}

\item{path}{path to download tiktok video to}
}
\description{
Function that enable to download tiktoks given a post_id
}
