% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{tk_info}
\alias{tk_info}
\title{tk_info}
\usage{
tk_info(scope, query, ...)
}
\arguments{
\item{scope}{Character indicating the endpoint to scrape (must be "hashtag", "user"
or "music")}

\item{query}{Character indicating the username/hashtag/music_id to scrape}
}
\description{
Function to get information on a given user/hashtag/piece of music
}
\examples{

\dontrun{
# Get info about a user:
tk_info(scope = "user", query = "willsmith")
# Get info about a hashtag
tk_info(scope = "hashtag", query = "cosplay")
}
}
