% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_hashtag_post}
\alias{get_hashtag_post}
\title{get_hashtag_post}
\usage{
get_hashtag_post(count, hashtag, ...)
}
\arguments{
\item{count}{Number of tiktoks to retrieve}

\item{hashtag}{Hashtag to scrape}
}
\description{
Function to get tiktoks from a given hashtag
}
