% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_hashtag_post}
\alias{get_hashtag_post}
\title{get_hashtag_post}
\usage{
get_hashtag_post(
  hashtag,
  n = 10000,
  cursor = 0,
  save_dir = NULL,
  port = NULL,
  ua = default_ua,
  ...
)
}
\arguments{
\item{hashtag}{Hashtag to scrape}

\item{n}{Number of tiktoks to retrieve}
}
\description{
Function to get tiktoks from a given hashtag
}
