% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_music_post}
\alias{get_music_post}
\title{get_music_post}
\usage{
get_music_post(music_id, n = 10000, port = NULL, ua = default_ua, ...)
}
\arguments{
\item{music_id}{Character indicating the \code{music_id} to scrape}

\item{n}{Numeric indicating the number of tiktoks to scrape}
}
\description{
Function to get tiktoks from a given music piece
}
