% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_music_post}
\alias{get_music_post}
\title{get_music_post}
\usage{
get_music_post(count, music_id, port = NULL, ...)
}
\arguments{
\item{count}{Numeric indicating the number of tiktoks to scrape}

\item{music_id}{Character indicating the \code{music_id} to scrape}
}
\description{
Function to get tiktoks from a given music piece
}
