% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_n}
\alias{get_n}
\title{get_n}
\usage{
get_n(
  type,
  n = 10000,
  cursor = 0,
  ua = default_ua,
  port = NULL,
  query_1 = NULL,
  query_2 = NULL,
  save_dir = NULL,
  query = NULL,
  vpn = F
)
}
\description{
Main function to get data from tiktok
}
