% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_n}
\alias{get_n}
\title{get_n}
\usage{
get_n(
  scope,
  n = 10000,
  start_date = lubridate::dmy("01-01-1900"),
  cursor = 0,
  query_1 = NULL,
  query_2 = "",
  save_dir = NULL,
  query = NULL,
  ...
)
}
\description{
Main function to get data from tiktok
}
