% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{get_url}
\alias{get_url}
\title{get_url}
\usage{
get_url(
  type,
  n = NULL,
  cursor = NULL,
  query_1 = NULL,
  query_2 = "",
  max = 0,
  min = 0
)
}
\description{
handle the api url based on a given type
}
