% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_user_post}
\alias{get_user_post}
\title{get_user_post}
\usage{
<<<<<<< HEAD
get_user_post(
  username,
  n = 10000,
  save_dir = NULL,
  port = NULL,
  ua = default_ua,
  ...
)
=======
get_user_post(count, username, ...)
>>>>>>> 16eb9ca88d97851895eba630e32e62f679ed8cf8
}
\arguments{
\item{username}{Character indicating the username to scrape}

\item{n}{Numeric indicating the number of tiktoks to scrape}
}
\description{
Function to get tiktoks from a given username
}
