% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_user_post}
\alias{get_user_post}
\title{get_user_post}
\usage{
get_user_post(count, username)
}
\arguments{
\item{count}{Numeric indicating the number of tiktoks to scrape}

\item{username}{Character indicating the username to scrape}
}
\description{
Function to get tiktoks from a given username
}
