% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tk_dl_video}
\alias{tk_dl_video}
\title{tk_dl_video}
\usage{
tk_dl_video(post_id = NULL, url = NULL, path, verbose = T, ...)
}
\arguments{
\item{url}{url of tiktok to scrape}

\item{path}{path to download tiktok video to}
}
\description{
Function that enable to download tiktoks
}
