% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{tk_info}
\alias{tk_info}
\title{tk_info}
\usage{
tk_info(scope, query, ua = default_ua, port = NULL, vpn = F)
}
\arguments{
\item{scope}{Character indicating the endpoint to scrape (must be "hashtag", "user"}

\item{query}{Character indicating the username/hashtag/music_id to scrape
or "music")}
}
\description{
Function to get information on a given user/hashtag/piece of music
}
