% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tk_posts}
\alias{tk_posts}
\title{tk_posts}
\usage{
tk_posts(
  scope,
  query,
  n = 10000,
  cursor = 0,
  save_dir = NULL,
  port = NULL,
  ua = default_ua
)
}
\arguments{
\item{scope}{Character indicating the endpoint to scrape (must be "user", "hashtag", "music" or "trends")}

\item{query}{Character indicating the username/hashtag/music_id to scrape}

\item{n}{Numeric indicating the number of tiktoks to scrape}
}
\description{
Get TikToks from a given username/hashtag or music_id
}
